Added 3 softwares.
The first program is finance_calculators.py, this program allows users to access two different financial calculators: an investment calculator and a home loan repayment calculator.
The second one is myfirstgame.py, this one is a basic game and it should be run alongside the 3 jpg images (monster, player, prize).
The third one is a task manager, it helps in managing tasks assigned to each member of a group. It should be run alongside the 5 text files and the setup.py file.